debug_screen.dartとmain.dartに変更を追加
